+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.NaturalTravel">
         <activity
-            android:name=".MainListaActivity"
+            android:name=".list.MainListaActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,7 +21,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".detalle.DetalleActivity"
             android:exported="true">
 
         </activity>
