--- app/src/main/java/com/TSRosies/naturaltravel/SitioNatural.kt
@@ -1,6 +1,0 @@
-package com.TSRosies.naturaltravel
-
-
-import com.google.gson.annotations.SerializedName
-
-class SitioNatural : ArrayList<SitioNaturalItem>()
