+++ app/src/main/java/com/TSRosies/naturaltravel/detalle/DetalleActivity.kt
@@ -1,18 +1,21 @@
-package com.TSRosies.naturaltravel
+package com.TSRosies.naturaltravel.detalle
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.TSRosies.naturaltravel.R
 import org.imaginativeworld.whynotimagecarousel.ImageCarousel
 import org.imaginativeworld.whynotimagecarousel.model.CarouselItem
 
-class MainActivity : AppCompatActivity() {
+class DetalleActivity : AppCompatActivity() {
     val list = mutableListOf<CarouselItem>()
     override fun onCreate(savedInstanceState: Bundle?) {
+
+
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         val carousel: ImageCarousel = findViewById(R.id.carousel)
         list.add(CarouselItem(imageDrawable = R.drawable.laguna_de_pedro_palo))
-        list.add(CarouselItem(imageDrawable = R.drawable.laguna_de_pedro_palo2 ))
+        list.add(CarouselItem(imageDrawable = R.drawable.laguna_de_pedro_palo2))
         carousel.addData(list)
     }
 }
