+++ app/src/main/java/com/TSRosies/naturaltravel/model/SitioNatural.kt
@@ -1,0 +1,4 @@
+package com.TSRosies.naturaltravel.model
+
+
+class SitioNatural : ArrayList<SitioNaturalItem>()
