+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
-    tools:context=".MainActivity">
+    tools:context=".detalle.DetalleActivity">
 
  <TextView
      android:id="@+id/textView"
